parts/overlay: make overlay compatible with darwin

+++ parts/overlays/fish.nix
@@ -1,10 +1,14 @@
 _: prev: {
-  fish = prev.symlinkJoin {
-    inherit (prev.fish) passthru;
-    name = "fish-nodesktop";
-    paths = [prev.fish];
-    postBuild = ''
-      rm $out/share/applications/fish.desktop
-    '';
-  };
+  fish =
+    if prev.stdenv.isLinux
+    then
+      prev.symlinkJoin {
+        inherit (prev.fish) passthru;
+        name = "fish-nodesktop";
+        paths = [prev.fish];
+        postBuild = ''
+          rm $out/share/applications/fish.desktop
+        '';
+      }
+    else prev.fish;
 }
