+++ parts/overlays/neovim.nix
@@ -1,10 +1,14 @@
 _: prev: {
-  neovim = prev.symlinkJoin {
-    inherit (prev.neovim) passthru;
-    name = "neovim-nodesktop";
-    paths = [prev.neovim];
-    postBuild = ''
-      rm $out/share/applications/*.desktop
-    '';
-  };
+  neovim =
+    if prev.stdenv.isLinux
+    then
+      prev.symlinkJoin {
+        inherit (prev.neovim) passthru;
+        name = "neovim-nodesktop";
+        paths = [prev.neovim];
+        postBuild = ''
+          rm $out/share/applications/*.desktop
+        '';
+      }
+    else prev.neovim;
 }
